requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
albumentations==1.0.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
numpy==1.23.5
numpy-quaternion==2022.4.3
matplotlib==3.5.1
pandas==1.4.3
scipy==1.10.1
scikit-learn==1.2.1
tqdm==4.62.3
tensorboard==2.13.*
PyYAML==6.0
torch==2.0.1+cu118
torchaudio==2.0.2
torchgeometry==0.1.2
torchvision==0.15.2+cu118
imageio==2.9.0
pycocotools==2.0.6